Extract both date formatters from the search filters into `DateFormatter` extensions.
